requirements.txt
numpy
pandas
scipy
requests
pycurl
certifi
beautifulsoup4
tabulate
openpyxl
tqdm
openai
